# backend/app/api/billing.py
"""Stripe billing: public plans, Checkout, Customer Portal, webhooks."""

# B2B note: this module covers candidate subscription Checkout + Customer Portal + webhooks only.
# A separate "per hire" / success-fee flow for employers (Stripe Connect, invoices, or usage-based)
# and a company billing dashboard are not implemented here — extend these handlers (or add a sibling
# router under the same Stripe config) once employer pricing models exist; avoid duplicating Stripe setup.

import logging
from typing import Annotated, Any

from fastapi import APIRouter, Depends, Header, HTTPException, Request, status
from sqlalchemy.orm import Session

import stripe
from app.config import Settings, get_settings
from app.core.deps import get_current_user
from app.database.models import User
from app.database.session import get_db
from app.schemas.billing import (
    CheckoutRequest,
    CheckoutResponse,
    PlanOut,
    PlansPublicResponse,
    PortalResponse,
)
from app.services import stripe_billing as stripe_svc

logger = logging.getLogger(__name__)

router = APIRouter()


def _stripe_object_to_dict(obj: object) -> dict[str, Any]:
    if isinstance(obj, dict):
        return obj
    to_dict = getattr(obj, "to_dict", None)
    if callable(to_dict):
        return to_dict()
    raise TypeError("Unsupported Stripe payload type")


def _checkout_configured(settings: Settings) -> bool:
    return bool(settings.stripe_secret_key and settings.stripe_price_id_premium)


def _plan_row(
    settings: Settings,
    *,
    id: str,
    name: str,
    description: str,
    max_tracked: int | None,
    stripe_price_configured: bool,
    price_id: str,
    fallback_usd: float,
) -> PlanOut:
    live: tuple[float, str] | None = None
    if stripe_price_configured and price_id.strip() and settings.stripe_secret_key.strip():
        live = stripe_svc.stripe_monthly_list_from_price(settings, price_id.strip())
    if live:
        amt, cur = live
        return PlanOut(
            id=id,
            name=name,
            description=description,
            max_tracked_applications=max_tracked,
            stripe_price_configured=stripe_price_configured,
            monthly_list_price_usd=stripe_svc.list_price_usd_hint(amt, cur),
            list_price_monthly=amt,
            list_price_currency=cur.upper(),
        )
    return PlanOut(
        id=id,
        name=name,
        description=description,
        max_tracked_applications=max_tracked,
        stripe_price_configured=stripe_price_configured,
        monthly_list_price_usd=fallback_usd,
        list_price_monthly=None,
        list_price_currency=None,
    )


@router.get("/plans", response_model=PlansPublicResponse)
def list_plans(settings: Annotated[Settings, Depends(get_settings)]) -> PlansPublicResponse:
    premium_ready = bool(settings.stripe_price_id_premium)
    pro_ready = bool(settings.stripe_price_id_pro)
    pm_types = stripe_svc.checkout_payment_method_types(settings)
    pm_note = stripe_svc.checkout_payment_methods_note(pm_types)
    return PlansPublicResponse(
        checkout_configured=_checkout_configured(settings),
        checkout_payment_methods=pm_types,
        payment_methods_note=pm_note,
        plans=[
            _plan_row(
                settings,
                id="free",
                name="Free",
                description="Core pipeline: discover roles, track applications, stay GDPR-first.",
                max_tracked=25,
                stripe_price_configured=False,
                price_id="",
                fallback_usd=0.0,
            ),
            _plan_row(
                settings,
                id="premium",
                name="Premium",
                description="Unlimited tracked applications, auto-apply where boards allow, priority roadmap.",
                max_tracked=None,
                stripe_price_configured=premium_ready,
                price_id=settings.stripe_price_id_premium,
                fallback_usd=4.99,
            ),
            _plan_row(
                settings,
                id="pro",
                name="Pro",
                description="Same entitlements as Premium today; reserved for team billing and higher limits.",
                max_tracked=None,
                stripe_price_configured=pro_ready,
                price_id=settings.stripe_price_id_pro,
                fallback_usd=9.99,
            ),
        ],
    )


@router.post("/checkout-session", response_model=CheckoutResponse)
def create_checkout_session(
    body: CheckoutRequest,
    user: Annotated[User, Depends(get_current_user)],
    settings: Annotated[Settings, Depends(get_settings)],
) -> CheckoutResponse:
    if not _checkout_configured(settings):
        raise HTTPException(
            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
            detail="Billing is not configured on this server.",
        )
    try:
        price_id = stripe_svc.price_id_for_plan(settings, body.plan)
    except ValueError as e:
        code = str(e)
        if code == "pro_not_configured":
            raise HTTPException(
                status_code=status.HTTP_400_BAD_REQUEST,
                detail="Pro plan is not available yet.",
            ) from e
        raise HTTPException(
            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
            detail="Billing is not configured on this server.",
        ) from e

    stripe_svc.configure_stripe(settings)
    try:
        md: dict[str, str] = {"user_id": str(user.id)}
        bn = (getattr(user, "billing_company_name", None) or "").strip()
        tid = (getattr(user, "billing_tax_id", None) or "").strip()
        if bn:
            md["billing_company_name"] = bn[:200]
        if tid:
            md["billing_tax_id"] = tid[:64]
        params: dict[str, object] = {
            "mode": "subscription",
            "line_items": [{"price": price_id, "quantity": 1}],
            "success_url": f"{settings.frontend_url}/dashboard/billing?checkout=success",
            "cancel_url": f"{settings.frontend_url}/dashboard/billing?checkout=cancel",
            "metadata": md,
            "subscription_data": {"metadata": md},
            "payment_method_types": stripe_svc.checkout_payment_method_types(settings),
            "allow_promotion_codes": True,
            "tax_id_collection": {"enabled": True},
        }
        if user.stripe_customer_id:
            params["customer"] = user.stripe_customer_id
        else:
            params["customer_email"] = user.email
        session = stripe.checkout.Session.create(**params)
    except stripe.StripeError as e:
        logger.warning("Stripe checkout failed: %s", e.user_message or str(e))
        raise HTTPException(
            status_code=status.HTTP_502_BAD_GATEWAY,
            detail="Payment provider error. Try again shortly.",
        ) from e

    url = session.get("url") if isinstance(session, dict) else session.url
    if not url:
        raise HTTPException(status_code=status.HTTP_502_BAD_GATEWAY, detail="Missing checkout URL.")
    return CheckoutResponse(url=url)


@router.post("/portal-session", response_model=PortalResponse)
def create_portal_session(
    user: Annotated[User, Depends(get_current_user)],
    settings: Annotated[Settings, Depends(get_settings)],
) -> PortalResponse:
    if not settings.stripe_secret_key:
        raise HTTPException(status_code=status.HTTP_503_SERVICE_UNAVAILABLE, detail="Billing not configured.")
    if not user.stripe_customer_id:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="No Stripe customer on file. Start a subscription from Checkout first.",
        )
    stripe_svc.configure_stripe(settings)
    try:
        portal = stripe.billing_portal.Session.create(
            customer=user.stripe_customer_id,
            return_url=f"{settings.frontend_url}/dashboard/billing",
        )
    except stripe.StripeError as e:
        logger.warning("Stripe portal failed: %s", e.user_message or str(e))
        raise HTTPException(
            status_code=status.HTTP_502_BAD_GATEWAY,
            detail="Payment provider error. Try again shortly.",
        ) from e

    url = portal.get("url") if isinstance(portal, dict) else portal.url
    if not url:
        raise HTTPException(status_code=status.HTTP_502_BAD_GATEWAY, detail="Missing portal URL.")
    return PortalResponse(url=url)


@router.post("/webhook")
async def stripe_webhook(
    request: Request,
    db: Annotated[Session, Depends(get_db)],
    settings: Annotated[Settings, Depends(get_settings)],
    stripe_signature: Annotated[str | None, Header(alias="stripe-signature")] = None,
) -> dict[str, str]:
    if not settings.stripe_webhook_secret or not settings.stripe_secret_key:
        raise HTTPException(status_code=status.HTTP_503_SERVICE_UNAVAILABLE, detail="Webhook not configured.")
    payload = await request.body()
    if not stripe_signature:
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Missing stripe-signature.")
    stripe_svc.configure_stripe(settings)
    try:
        event = stripe.Webhook.construct_event(
            payload=payload,
            sig_header=stripe_signature,
            secret=settings.stripe_webhook_secret,
        )
    except ValueError as e:
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid payload.") from e
    except stripe.SignatureVerificationError as e:
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid signature.") from e

    event_d = _stripe_object_to_dict(event)
    etype = event_d["type"]
    obj_d = _stripe_object_to_dict(event_d["data"]["object"])

    try:
        if etype == "checkout.session.completed":
            stripe_svc.process_checkout_completed(db, obj_d, settings)
        elif etype == "customer.subscription.updated":
            stripe_svc.process_subscription_updated(db, obj_d, settings)
        elif etype == "customer.subscription.deleted":
            stripe_svc.process_subscription_deleted(db, obj_d)
        elif etype == "invoice.payment_succeeded":
            stripe_svc.process_invoice_payment_succeeded(db, obj_d, settings)
        elif etype == "invoice.payment_failed":
            stripe_svc.process_invoice_payment_failed(db, obj_d, settings)
        elif etype == "checkout.session.async_payment_failed":
            stripe_svc.process_checkout_session_async_payment_failed(db, obj_d, settings)
    except Exception:
        logger.exception("Stripe webhook handler failed for %s", etype)
        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Webhook handler error.")

    return {"received": "true"}